requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gevent==20.9.0
greenlet==0.4.17
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
prettytable==1.0.1
pycparser==2.20
pyecharts==1.9.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
websocket==0.2.1
websocket-client==0.57.0
zope.event==4.5.0
zope.interface==5.1.2
